Fixed bug on MacOSX. When right clicking on any element of xfm, ccfm or vdm, eclipse gets frozen.
We are not sure about it, but probably pure variants starts loading all the files inside output/ directory for internal checkings. As we have a node_modules folder with more than 500MB and 30K files, it tries to analyze all of them. The solution adopted was to move node_modules folder outside the project scope (it is now created in the workspace folder) and links to node_modules folder inside the generated product is temporal, when gulp default is done, it is removed too.